requirements.txt
matplotlib==3.7.2
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
streamlit